refactor: move stream out of use for now
Not currently used and requires a complete rewrite
