+++ src/chunk.rs
@@ -61,12 +61,3 @@ pub(crate) fn batch_chunks(bytes: Bytes) -> (usize, Vec<EncryptionBatch>) {
 
     (num_chunks, batches)
 }
-
-/// Calculate (start_position, end_position) for each chunk for the input file size
-pub(crate) fn batch_positions(data_size: usize) -> Vec<(usize, usize)> {
-    let num_chunks = get_num_chunks(data_size);
-
-    (0..num_chunks)
-        .map(|index| get_start_end_positions(data_size, index))
-        .collect()
-}
